remove ggml kernel usage in cpu backend
