+++ runtime/onert/backend/cpu/CMakeLists.txt
@@ -1,7 +1,7 @@
 nnfw_find_package(Ruy REQUIRED)
 
 file(GLOB SOURCES "*.cc")
-list(APPEND SOURCES ops/GGMLHelper.cc ops/OperationUtils.cc)
+list(APPEND SOURCES ops/OperationUtils.cc)
 macro(OP NAME)
   list(APPEND SOURCES ops/${NAME}Layer.cc)
 endmacro(OP)
@@ -20,8 +20,6 @@ target_link_libraries(${LIB_ONERT_BACKEND_CPU} PRIVATE nnfw_coverage)
 target_link_libraries(${LIB_ONERT_BACKEND_CPU} PRIVATE ruy)
 target_link_libraries(${LIB_ONERT_BACKEND_CPU} INTERFACE ruy_instrumentation)
 target_link_libraries(${LIB_ONERT_BACKEND_CPU} PRIVATE ndarray)
-# Set public: ExternalContext is used in train backend
-target_link_libraries(${LIB_ONERT_BACKEND_CPU} PUBLIC ggml)
 
 set_target_properties(${LIB_ONERT_BACKEND_CPU} PROPERTIES
   OUTPUT_NAME backend_cpu
