+++ runtime/onert/backend/cpu/ExternalContext.h
@@ -19,7 +19,6 @@
 
 #include <util/ConfigSource.h>
 #include <ruy/context.h>
-#include <ggml.h>
 
 #include <memory>
 
@@ -36,14 +35,11 @@ public:
 
   int32_t maxNumThreads() const { return _max_num_threads; }
 
-  void initGgmlContext();
-
   ruy::Context *ruy_context() const { return _ruy_context.get(); }
 
 private:
   int32_t _max_num_threads;
   const std::unique_ptr<ruy::Context> _ruy_context;
-  std::unique_ptr<ggml_context, decltype(&ggml_free)> _ggml_context{nullptr, &ggml_free};
 };
 
 } // namespace onert::backend::cpu
