+++ runtime/onert/backend/cpu/ops/FullyConnectedLayer.cc
@@ -16,7 +16,6 @@
 
 #include "FullyConnectedLayer.h"
 
-#include "GGMLHelper.h"
 #include "../Tensor.h"
 #include "../KernelGenerator.h"
 #include "../Validator.h"
@@ -28,7 +27,21 @@
 namespace onert::backend::cpu
 {
 
-void Validator::visit(const ir::operation::FullyConnected &) { _supported = true; }
+void Validator::visit(const ir::operation::FullyConnected &node)
+{
+  using ir::operation::FullyConnected;
+
+  const auto weight_index{node.getInputs().at(FullyConnected::Input::WEIGHT)};
+  const auto weight_node = &_graph.operands().at(weight_index);
+
+  _supported = false;
+
+  if (weight_node->typeInfo().type() == ir::DataType::QUANT_GGML_Q4_0 ||
+      weight_node->typeInfo().type() == ir::DataType::QUANT_GGML_Q8_0)
+    return;
+
+  _supported = true;
+}
 
 void KernelGenerator::visit(const ir::operation::FullyConnected &node)
 {
@@ -206,39 +219,6 @@ void FullyConnectedLayer::fullyConnectedSparseWeight()
     throw std::runtime_error{"FullyConnected: unsupported sparsity"};
 }
 
-void FullyConnectedLayer::fullyConnectedGGMLWeight()
-{
-  if (_bias)
-    throw std::runtime_error{"FullyConnected: GGML weights format does not support bias yet."};
-
-  // convert tensor
-  auto input = getGGMLTensor(_input);
-  auto weights = getGGMLTensor(_weights);
-  auto output = getGGMLTensor(_output);
-  {
-    output.op = GGML_OP_MUL_MAT;
-    output.src[0] = &weights;
-    output.src[1] = &input;
-  }
-  auto *nodes = &output;
-
-  // create graph
-  struct ggml_cgraph graph;
-  {
-    memset(&graph, 0, sizeof(graph));
-    graph.n_nodes = 1;
-    graph.nodes = &nodes;
-  }
-
-  // get cplan
-  auto cplan = ggml_graph_plan(&graph, _external_context->maxNumThreads());
-  std::vector<uint8_t> buf(cplan.work_size);
-  cplan.work_data = buf.data();
-
-  // compute
-  ggml_graph_compute(&graph, &cplan);
-}
-
 void FullyConnectedLayer::fullyConnected16x1Float32()
 {
 #if defined(__aarch64__) && defined(USE_NEON)
@@ -279,10 +259,6 @@ void FullyConnectedLayer::configure(const IPortableTensor *input, const IPortabl
   }
 #endif
   _external_context = external_context;
-
-  if (_weights->data_type() == OperandType::QUANT_GGML_Q4_0 ||
-      _weights->data_type() == OperandType::QUANT_GGML_Q8_0)
-    _external_context->initGgmlContext();
 }
 
 void FullyConnectedLayer::run()
@@ -294,11 +270,6 @@ void FullyConnectedLayer::run()
   else if (_weights->sparsity())
   {
     fullyConnectedSparseWeight();
-  }
-  else if (_weights->data_type() == OperandType::QUANT_GGML_Q4_0 ||
-           _weights->data_type() == OperandType::QUANT_GGML_Q8_0)
-  {
-    fullyConnectedGGMLWeight();
   }
   else if (_input->data_type() == OperandType::FLOAT32)
   {
