--- runtime/onert/backend/cpu/ops/GGMLHelper.h
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2024 Samsung Electronics Co., Ltd. All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in riting, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef __ONERT_BACKEND_CPU_GGML_HELPER_H__
-#define __ONERT_BACKEND_CPU_GGML_HELPER_H__
-
-#include <backend/IPortableTensor.h>
-
-#include <ggml.h>
-
-namespace onert::backend::cpu::ops
-{
-
-struct ggml_tensor getGGMLTensor(const IPortableTensor *tensor);
-
-} // namespace onert::backend::cpu::ops
-
-#endif
