+++ runtime/onert/backend/cpu/ops/GatherLayer.cc
@@ -17,7 +17,6 @@
 #include "GatherLayer.h"
 
 #include "OperationUtils.h"
-#include "GGMLHelper.h"
 #include "../KernelGenerator.h"
 #include "../Validator.h"
 
@@ -26,7 +25,20 @@
 namespace onert::backend::cpu
 {
 
-void Validator::visit(const ir::operation::Gather &) { _supported = true; }
+void Validator::visit(const ir::operation::Gather &node)
+{
+  using ir::operation::Gather;
+
+  const auto input_index{node.getInputs().at(Gather::Input::INPUT)};
+  const auto input_node = &_graph.operands().at(input_index);
+
+  _supported = false;
+
+  if (input_node->typeInfo().type() == ir::DataType::QUANT_GGML_Q4_0)
+    return;
+
+  _supported = true;
+}
 
 void KernelGenerator::visit(const ir::operation::Gather &node)
 {
@@ -43,7 +55,7 @@ void KernelGenerator::visit(const ir::operation::Gather &node)
 
   auto fn = std::make_unique<ops::GatherLayer>();
 
-  fn->configure(input_tensor, indices_tensor, output_tensor, axis, _external_context.get());
+  fn->configure(input_tensor, indices_tensor, output_tensor, axis);
 
   _return_fn = std::move(fn);
 }
@@ -54,16 +66,12 @@ namespace onert::backend::cpu::ops
 {
 
 void GatherLayer::configure(const IPortableTensor *input, const IPortableTensor *indices,
-                            IPortableTensor *output, int32_t axis, ExternalContext *ctx)
+                            IPortableTensor *output, int32_t axis)
 {
   _input = input;
   _indices = indices;
   _axis = axis;
   _output = output;
-  _ctx = ctx;
-
-  if (_input->data_type() == OperandType::QUANT_GGML_Q4_0)
-    ctx->initGgmlContext();
 }
 
 template <typename InputType> void GatherLayer::runByInputType()
@@ -97,53 +105,6 @@ template <typename InputType> void GatherLayer::runByInputType()
   }
 }
 
-void GatherLayer::runByGGMLQuantInputType()
-{
-  // Supporting condition
-  // Input: rank 2
-  // Indice: rank < 4 or rank 4 with dim(0) = 1, INT32
-  // Axis: 0
-  if (getShape(_input).DimensionsCount() != 2)
-    throw std::runtime_error("Gather: block quantized input tensor must be rank 2");
-
-  if (getShape(_indices).DimensionsCount() >= 4 &&
-      (getShape(_indices).DimensionsCount() != 4 || getShape(_indices).Dims(0) != 1))
-    throw std::runtime_error("Gather: invalid indices tensor shape");
-
-  if (_indices->data_type() != ir::DataType::INT32)
-    throw std::runtime_error("Gather: indices tensor must be int32 type");
-
-  if (_axis != 0)
-    throw std::runtime_error("Gather: axis must be 0");
-
-  // convert tensor
-  auto input = getGGMLTensor(_input);
-  auto indices = getGGMLTensor(_indices);
-  auto output = getGGMLTensor(_output);
-  {
-    output.op = GGML_OP_GET_ROWS;
-    output.src[0] = &input;
-    output.src[1] = &indices;
-  }
-  auto *nodes = &output;
-
-  // create graph
-  struct ggml_cgraph graph;
-  {
-    memset(&graph, 0, sizeof(graph));
-    graph.n_nodes = 1;
-    graph.nodes = &nodes;
-  }
-
-  // get cplan
-  auto cplan = ggml_graph_plan(&graph, _ctx->maxNumThreads());
-  std::vector<uint8_t> buf(cplan.work_size);
-  cplan.work_data = buf.data();
-
-  // compute
-  ggml_graph_compute(&graph, &cplan);
-}
-
 void GatherLayer::run()
 {
   switch (_input->data_type())
@@ -156,9 +117,6 @@ void GatherLayer::run()
       break;
     case OperandType::INT32:
       runByInputType<int32_t>();
-      break;
-    case OperandType::QUANT_GGML_Q4_0:
-      runByGGMLQuantInputType();
       break;
     case OperandType::BOOL8:
       runByInputType<bool>();
