+++ runtime/onert/backend/cpu/ops/GatherLayer.h
@@ -17,8 +17,6 @@
 #ifndef __ONERT_BACKEND_CPU_OPS_GATHERLAYER_H__
 #define __ONERT_BACKEND_CPU_OPS_GATHERLAYER_H__
 
-#include "../ExternalContext.h"
-
 #include <backend/IPortableTensor.h>
 
 #include <exec/IFunction.h>
@@ -29,20 +27,19 @@ namespace onert::backend::cpu::ops
 class GatherLayer : public ::onert::exec::IFunction
 {
 public:
-  GatherLayer() : _input{nullptr}, _indices{nullptr}, _output{nullptr}, _axis{-1}, _ctx{nullptr}
+  GatherLayer() : _input{nullptr}, _indices{nullptr}, _output{nullptr}, _axis{-1}
   {
     // DO NOTHING
   }
 
 public:
   void configure(const IPortableTensor *input, const IPortableTensor *indices,
-                 IPortableTensor *output, int32_t axis, ExternalContext *ctx);
+                 IPortableTensor *output, int32_t axis);
 
   void run() override;
 
 private:
   template <typename OpType> void runByInputType();
-  void runByGGMLQuantInputType();
 
 private:
   const IPortableTensor *_input;
@@ -50,7 +47,6 @@ private:
   IPortableTensor *_output;
 
   int32_t _axis;
-  ExternalContext *_ctx;
 };
 
 } // namespace onert::backend::cpu::ops
