+++ runtime/onert/backend/ggml/CMakeLists.txt
@@ -14,6 +14,10 @@ target_link_libraries(${LIB_ONERT_BACKEND_GGML} PRIVATE nnfw_common)
 target_link_libraries(${LIB_ONERT_BACKEND_GGML} PRIVATE nnfw_coverage)
 target_link_libraries(${LIB_ONERT_BACKEND_GGML} PRIVATE ggml)
 
+# Some used functions makes deprecated declarations warning
+# TODO Resolve workaround
+target_compile_options(${LIB_ONERT_BACKEND_GGML} PRIVATE -Wno-deprecated-declarations)
+
 set_target_properties(${LIB_ONERT_BACKEND_GGML} PROPERTIES
   OUTPUT_NAME backend_ggml
   INSTALL_RPATH ${ONERT_RPATH_PLUGIN})
